Remove old fieldfield.vue component and replace with new one update any references.
