+++ boranga/frontend/boranga/src/components/internal/conservation_status/amendment_request.vue
@@ -50,7 +50,7 @@
 import Vue from 'vue'
 import modal from '@vue-utils/bootstrap-modal.vue'
 import alert from '@vue-utils/alert.vue'
-import FileField2 from '@/components/forms/filefield2.vue'
+import FileField2 from '@/components/forms/filefield.vue'
 
 import { helpers, api_endpoints } from "@/utils/hooks.js"
 export default {
