+++ boranga/frontend/boranga/src/components/internal/conservation_status/proposed_issuance.vue
@@ -79,7 +79,7 @@
 
 import modal from '@vue-utils/bootstrap-modal.vue'
 import alert from '@vue-utils/alert.vue'
-import FileField2 from '@/components/forms/filefield2.vue'
+import FileField2 from '@/components/forms/filefield.vue'
 import { helpers, api_endpoints } from "@/utils/hooks.js"
 export default {
     name: 'Proposed-Approval',
