+++ boranga/frontend/boranga/src/components/internal/species_communities/proposal_add_requirement.vue
@@ -151,7 +151,7 @@
 </template>
 
 <script>
-import FileField2 from '@/components/forms/filefield2.vue'
+import FileField2 from '@/components/forms/filefield.vue'
 import modal from '@vue-utils/bootstrap-modal.vue'
 import alert from '@vue-utils/alert.vue'
 import { helpers, api_endpoints } from "@/utils/hooks.js"
