+++ boranga/frontend/boranga/src/components/internal/species_communities/species_rename.vue
@@ -31,7 +31,7 @@
 <script>
 import modal from '@vue-utils/bootstrap-modal.vue'
 import alert from '@vue-utils/alert.vue'
-import FileField2 from '@/components/forms/filefield2.vue'
+import FileField2 from '@/components/forms/filefield.vue'
 import SpeciesCommunitiesComponent from '@/components/form_species_communities.vue'
 import { helpers, api_endpoints } from "@/utils/hooks.js"
 export default {
